refactor: move divide_viewport to its own file
