+++ src/main.rs
@@ -44,6 +44,7 @@ pub(crate) mod ui_tree;
 pub(crate) mod undo_tree;
 mod utils;
 
+mod divide_viewport;
 mod env;
 use std::{rc::Rc, sync::Mutex};
